# backend/blockchain_server/openapi.yaml
openapi: 3.0.0
info:
  title: Blockchain Transactions API
  version: 1.0.0
paths:
  /transactions:
    get:
      summary: Retrieve all transactions in the pool
      description: Returns the list of pending transactions in the transaction pool.
      operationId: getTransactions
      responses:
        '200':
          description: A list of pending transactions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    description: The list of pending transactions.
                    items:
                      $ref: '#/components/schemas/Transaction'
                  length:
                    type: integer
                    description: The total number of pending transactions.
        '400':
          description: Bad Request. Invalid method or error occurred.
    post:
      summary: Create a new transaction
      description: Submits a new transaction to the transaction pool. The request requires a valid signature and public key for verification.
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
      responses:
        '201':
          description: Transaction created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
        '400':
          description: Bad Request. Transaction creation failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
        '422':
          description: Validation failed. Missing or invalid fields in the request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  error:
                    type: string
                    example: "Missing sender_blockchain_address"
components:
  schemas:
    Transaction:
      type: object
      properties:
        sender_blockchain_address:
          type: string
          description: The blockchain address of the sender.
        recipient_blockchain_address:
          type: string
          description: The blockchain address of the recipient.
        value:
          type: number
          format: float
          description: The value of the transaction.
    TransactionRequest:
      type: object
      properties:
        sender_blockchain_address:
          type: string
          description: The blockchain address of the sender.
          example: "address1"
        recipient_blockchain_address:
          type: string
          description: The blockchain address of the recipient.
          example: "address2"
        sender_public_key:
          type: string
          description: The sender's public key, used for signature verification.
          example: "public_key1"
        value:
          type: number
          format: float
          description: The value of the transaction.
          example: 50.0
        signature:
          type: string
          description: The transaction signature.
          example: "signature1"